9180 Remove formatters package from specs and rename/ move files
