+++ app/uk/gov/hmrc/tai/model/domain/TaxOnOtherIncome.scala
@@ -27,51 +27,45 @@ case class TaxOnOtherIncome(tax: BigDecimal)
 object TaxOnOtherIncome {
   private val NonCodedIncome = 19
 
-  val taxOnOtherIncomeRead = new Reads[Option[BigDecimal]] {
-    override def reads(json: JsValue): JsResult[Option[BigDecimal]] =
-      JsSuccess(json.as[Option[TaxOnOtherIncome]](taxOnOtherIncomeReads) map (_.tax))
+  val taxOnOtherIncomeRead: Reads[Option[BigDecimal]] = (json: JsValue) =>
+    JsSuccess(json.as[Option[TaxOnOtherIncome]](taxOnOtherIncomeReads) map (_.tax))
+
+  val taxAccountSummaryReads: Reads[BigDecimal] = (json: JsValue) => {
+    val taxOnOtherIncome =
+      json.as[Option[TaxOnOtherIncome]](taxOnOtherIncomeReads) map (_.tax) getOrElse BigDecimal(0)
+    val totalLiabilityTax = (json \ "totalLiability" \ "totalLiability").asOpt[BigDecimal].getOrElse(BigDecimal(0))
+
+    JsSuccess(totalLiabilityTax - taxOnOtherIncome)
   }
 
-  val taxAccountSummaryReads = new Reads[BigDecimal] {
-    override def reads(json: JsValue): JsResult[BigDecimal] = {
-      val taxOnOtherIncome =
-        json.as[Option[TaxOnOtherIncome]](taxOnOtherIncomeReads) map (_.tax) getOrElse BigDecimal(0)
-      val totalLiabilityTax = (json \ "totalLiability" \ "totalLiability").asOpt[BigDecimal].getOrElse(BigDecimal(0))
+  val taxOnOtherIncomeReads: Reads[Option[TaxOnOtherIncome]] = (json: JsValue) => {
+    val iabdSummaries = totalLiabilityIabds(json, "totalIncome", Seq("nonSavings"))
+    val nonCodedIncomeAmount = iabdSummaries.find(_.componentType == NonCodedIncome).map(_.amount)
 
-      JsSuccess(totalLiabilityTax - taxOnOtherIncome)
-    }
-  }
-
-  val taxOnOtherIncomeReads = new Reads[Option[TaxOnOtherIncome]] {
-    override def reads(json: JsValue): JsResult[Option[TaxOnOtherIncome]] = {
-      val iabdSummaries = totalLiabilityIabds(json, "totalIncome", Seq("nonSavings"))
-      val nonCodedIncomeAmount = iabdSummaries.find(_.componentType == NonCodedIncome).map(_.amount)
-
-      @tailrec
-      def calculateTaxOnOtherIncome(
-        incomeAndRateBands: Seq[RateBand],
-        nonCodedIncome: BigDecimal,
-        total: BigDecimal = 0
-      ): BigDecimal =
-        incomeAndRateBands match {
-          case Nil => total
-          case xs if nonCodedIncome > xs.head.income =>
-            val newTotal = xs.head.income * (xs.head.rate / 100)
-            calculateTaxOnOtherIncome(xs.tail, nonCodedIncome - xs.head.income, total + newTotal)
-          case xs if nonCodedIncome <= xs.head.income =>
-            val newTotal = nonCodedIncome * (xs.head.rate / 100)
-            total + newTotal
-          case _ => throw new RuntimeException("Incorrect rate band")
-        }
-
-      (nonCodedIncomeAmount, incomeAndRateBands(json)) match {
-        case (None, _)      => JsSuccess(None)
-        case (Some(_), Nil) => JsSuccess(None)
-        case (Some(amount), incomeAndRateBands) =>
-          val remainingTaxOnOtherIncome = calculateTaxOnOtherIncome(incomeAndRateBands, amount)
-          JsSuccess(Some(TaxOnOtherIncome(remainingTaxOnOtherIncome)))
+    @tailrec
+    def calculateTaxOnOtherIncome(
+      incomeAndRateBands: Seq[RateBand],
+      nonCodedIncome: BigDecimal,
+      total: BigDecimal = 0
+    ): BigDecimal =
+      incomeAndRateBands match {
+        case Nil => total
+        case xs if nonCodedIncome > xs.head.income =>
+          val newTotal = xs.head.income * (xs.head.rate / 100)
+          calculateTaxOnOtherIncome(xs.tail, nonCodedIncome - xs.head.income, total + newTotal)
+        case xs if nonCodedIncome <= xs.head.income =>
+          val newTotal = nonCodedIncome * (xs.head.rate / 100)
+          total + newTotal
+        case _ => throw new RuntimeException("Incorrect rate band")
       }
 
+    (nonCodedIncomeAmount, incomeAndRateBands(json)) match {
+      case (None, _)      => JsSuccess(None)
+      case (Some(_), Nil) => JsSuccess(None)
+      case (Some(amount), incomeAndRateBands) =>
+        val remainingTaxOnOtherIncome = calculateTaxOnOtherIncome(incomeAndRateBands, amount)
+        JsSuccess(Some(TaxOnOtherIncome(remainingTaxOnOtherIncome)))
     }
+
   }
 }
