+++ test/uk/gov/hmrc/tai/model/domain/TaxComponentTypeSpec.scala
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2023 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.tai.model.domain
+
+import org.mockito.MockitoSugar
+import org.scalatestplus.play.PlaySpec
+import play.api.libs.json.{JsString, Json}
+import uk.gov.hmrc.tai.model.domain.TaxComponentType.codingComponentTypeWrites
+
+class TaxComponentTypeSpec extends PlaySpec with MockitoSugar {
+  "codingComponentTypeWrites" must {
+    "write tax component type correctly to json" when {
+      "tax component type is having valid type" in {
+        Json.toJson(GiftAidPayments)(codingComponentTypeWrites.writes(_)) mustBe JsString("GiftAidPayments")
+      }
+    }
+  }
+}
