+++ test/uk/gov/hmrc/tai/model/domain/TaxOnOtherIncomeSpec.scala
@@ -1,0 +1,478 @@
+/*
+ * Copyright 2023 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.tai.model.domain
+
+import org.mockito.MockitoSugar
+import org.scalatestplus.play.PlaySpec
+import play.api.libs.json.{JsArray, JsNull, Json}
+import uk.gov.hmrc.tai.model.domain.TaxOnOtherIncome.{taxAccountSummaryReads, taxOnOtherIncomeRead, taxOnOtherIncomeReads}
+
+class TaxOnOtherIncomeSpec extends PlaySpec with MockitoSugar {
+
+  "taxAccountSummaryReads" must {
+    "return the totalEstTax from the hods response" when {
+      "totalLiability val is present in totalLiability section" in {
+        val json = Json.obj(
+          "totalLiability" -> Json.obj(
+            "totalLiability" -> 1234.56
+          )
+        )
+
+        json.as[BigDecimal](taxAccountSummaryReads) mustBe BigDecimal(1234.56)
+      }
+    }
+    "return zero totalEstTax" when {
+      "totalLiability section is NOT present" in {
+        val json = Json.obj()
+        json.as[BigDecimal](taxAccountSummaryReads) mustBe BigDecimal(0)
+      }
+      "totalLiability section is null" in {
+        val json = Json.obj(
+          "totalLiability" -> JsNull
+        )
+        json.as[BigDecimal](taxAccountSummaryReads) mustBe BigDecimal(0)
+      }
+      "totalLiability section is present but the totalLiability value is not present inside" in {
+        val json = Json.obj(
+          "totalLiability" -> Json.obj()
+        )
+        json.as[BigDecimal](taxAccountSummaryReads) mustBe BigDecimal(0)
+      }
+      "totalLiability section is present but the totalLiability value is null inside" in {
+        val json = Json.obj(
+          "totalLiability" -> JsNull
+        )
+        json.as[BigDecimal](taxAccountSummaryReads) mustBe BigDecimal(0)
+      }
+    }
+
+    "return totalEstTax" when {
+      "tax on other income is present" in {
+        val json = Json.obj(
+          "totalLiability" -> Json.obj(
+            "totalLiability" -> 1234.56,
+            "nonSavings" -> Json.obj(
+              "totalIncome" -> Json.obj(
+                "iabdSummaries" -> JsArray(
+                  Seq(
+                    Json.obj(
+                      "amount"         -> 100,
+                      "type"           -> 19,
+                      "npsDescription" -> "Non-Coded Income",
+                      "employmentId"   -> JsNull
+                    ),
+                    Json.obj(
+                      "amount"         -> 100,
+                      "type"           -> 84,
+                      "npsDescription" -> "Job-Seeker Allowance",
+                      "employmentId"   -> JsNull
+                    )
+                  )
+                )
+              ),
+              "taxBands" -> JsArray(
+                Seq(
+                  Json.obj(
+                    "bandType" -> "B",
+                    "income"   -> 1000,
+                    "taxCode"  -> "BR",
+                    "rate"     -> 40
+                  ),
+                  Json.obj(
+                    "bandType" -> "D0",
+                    "taxCode"  -> "BR",
+                    "income"   -> 1000,
+                    "rate"     -> 20
+                  )
+                )
+              )
+            )
+          )
+        )
+
+        json.as[BigDecimal](taxAccountSummaryReads) mustBe BigDecimal(1194.56)
+      }
+
+    }
+  }
+
+  "TaxOnOtherIncomeFormatters" must {
+    "return tax on other income" when {
+      "non-coded income is present and highest rate is 40%" in {
+        val json = Json.obj(
+          "totalLiability" -> Json.obj(
+            "nonSavings" -> Json.obj(
+              "totalIncome" -> Json.obj(
+                "iabdSummaries" -> JsArray(
+                  Seq(
+                    Json.obj(
+                      "amount"         -> 100,
+                      "type"           -> 19,
+                      "npsDescription" -> "Non-Coded Income",
+                      "employmentId"   -> JsNull
+                    ),
+                    Json.obj(
+                      "amount"         -> 100,
+                      "type"           -> 84,
+                      "npsDescription" -> "Job-Seeker Allowance",
+                      "employmentId"   -> JsNull
+                    )
+                  )
+                )
+              ),
+              "taxBands" -> JsArray(
+                Seq(
+                  Json.obj(
+                    "bandType" -> "B",
+                    "income"   -> 1000,
+                    "taxCode"  -> "BR",
+                    "rate"     -> 40
+                  ),
+                  Json.obj(
+                    "bandType" -> "D0",
+                    "taxCode"  -> "BR",
+                    "income"   -> 1000,
+                    "rate"     -> 20
+                  )
+                )
+              )
+            )
+          )
+        )
+
+        json.as[Option[TaxOnOtherIncome]](taxOnOtherIncomeReads) mustBe Some(TaxOnOtherIncome(40))
+      }
+
+      "non-coded income is present and equal to highest rate income " in {
+        val json = Json.obj(
+          "totalLiability" -> Json.obj(
+            "nonSavings" -> Json.obj(
+              "totalIncome" -> Json.obj(
+                "iabdSummaries" -> JsArray(
+                  Seq(
+                    Json.obj(
+                      "amount"         -> 1000,
+                      "type"           -> 19,
+                      "npsDescription" -> "Non-Coded Income",
+                      "employmentId"   -> JsNull
+                    ),
+                    Json.obj(
+                      "amount"         -> 100,
+                      "type"           -> 84,
+                      "npsDescription" -> "Job-Seeker Allowance",
+                      "employmentId"   -> JsNull
+                    )
+                  )
+                )
+              ),
+              "taxBands" -> JsArray(
+                Seq(
+                  Json.obj(
+                    "bandType" -> "B",
+                    "income"   -> 1000,
+                    "taxCode"  -> "BR",
+                    "rate"     -> 40
+                  ),
+                  Json.obj(
+                    "bandType" -> "D0",
+                    "taxCode"  -> "BR",
+                    "income"   -> 1000,
+                    "rate"     -> 20
+                  )
+                )
+              )
+            )
+          )
+        )
+
+        json.as[Option[TaxOnOtherIncome]](taxOnOtherIncomeReads) mustBe Some(TaxOnOtherIncome(400))
+      }
+
+      "non-coded income is present and scattered in multiple rate bands " in {
+        val json = Json.obj(
+          "totalLiability" -> Json.obj(
+            "nonSavings" -> Json.obj(
+              "totalIncome" -> Json.obj(
+                "iabdSummaries" -> JsArray(
+                  Seq(
+                    Json.obj(
+                      "amount"         -> 10000,
+                      "type"           -> 19,
+                      "npsDescription" -> "Non-Coded Income",
+                      "employmentId"   -> JsNull
+                    ),
+                    Json.obj(
+                      "amount"         -> 100,
+                      "type"           -> 84,
+                      "npsDescription" -> "Job-Seeker Allowance",
+                      "employmentId"   -> JsNull
+                    )
+                  )
+                )
+              ),
+              "taxBands" -> JsArray(
+                Seq(
+                  Json.obj(
+                    "bandType" -> "B",
+                    "income"   -> 5000,
+                    "taxCode"  -> "BR",
+                    "rate"     -> 40
+                  ),
+                  Json.obj(
+                    "bandType" -> "D0",
+                    "taxCode"  -> "BR",
+                    "income"   -> 1000,
+                    "rate"     -> 20
+                  ),
+                  Json.obj(
+                    "bandType" -> "D0",
+                    "taxCode"  -> "BR",
+                    "income"   -> 8000,
+                    "rate"     -> 10
+                  )
+                )
+              )
+            )
+          )
+        )
+
+        json.as[Option[TaxOnOtherIncome]](taxOnOtherIncomeReads) mustBe Some(TaxOnOtherIncome(2600))
+      }
+
+      "non-coded income is present and highest rate is 20%" in {
+        val json = Json.obj(
+          "totalLiability" -> Json.obj(
+            "nonSavings" -> Json.obj(
+              "totalIncome" -> Json.obj(
+                "iabdSummaries" -> JsArray(
+                  Seq(
+                    Json.obj(
+                      "amount"         -> 100,
+                      "type"           -> 19,
+                      "npsDescription" -> "Non-Coded Income",
+                      "employmentId"   -> JsNull
+                    ),
+                    Json.obj(
+                      "amount"         -> 100,
+                      "type"           -> 84,
+                      "npsDescription" -> "Job-Seeker Allowance",
+                      "employmentId"   -> JsNull
+                    )
+                  )
+                )
+              ),
+              "taxBands" -> JsArray(
+                Seq(
+                  Json.obj(
+                    "bandType" -> "B",
+                    "income"   -> 1000,
+                    "taxCode"  -> "BR",
+                    "rate"     -> 20
+                  )
+                )
+              )
+            )
+          )
+        )
+
+        json.as[Option[TaxOnOtherIncome]](taxOnOtherIncomeReads) mustBe Some(TaxOnOtherIncome(20))
+      }
+
+      "non-coded income is not present" in {
+        val json = Json.obj(
+          "totalLiability" -> Json.obj(
+            "nonSavings" -> Json.obj(
+              "totalIncome" -> Json.obj(
+                "iabdSummaries" -> JsArray(
+                  Seq(
+                    Json.obj(
+                      "amount"         -> 100,
+                      "type"           -> 84,
+                      "npsDescription" -> "Job-Seeker Allowance",
+                      "employmentId"   -> JsNull
+                    )
+                  )
+                )
+              ),
+              "taxBands" -> JsArray(
+                Seq(
+                  Json.obj(
+                    "bandType" -> "B",
+                    "income"   -> 1000,
+                    "taxCode"  -> "BR",
+                    "rate"     -> 20
+                  )
+                )
+              )
+            )
+          )
+        )
+
+        json.as[Option[TaxOnOtherIncome]](taxOnOtherIncomeReads) mustBe None
+      }
+
+      "non-coded income is present and tax bands are not present" in {
+        val json = Json.obj(
+          "totalLiability" -> Json.obj(
+            "nonSavings" -> Json.obj(
+              "totalIncome" -> Json.obj(
+                "iabdSummaries" -> JsArray(
+                  Seq(
+                    Json.obj(
+                      "amount"         -> 100,
+                      "type"           -> 19,
+                      "npsDescription" -> "Non-Coded Income",
+                      "employmentId"   -> JsNull
+                    ),
+                    Json.obj(
+                      "amount"         -> 100,
+                      "type"           -> 84,
+                      "npsDescription" -> "Job-Seeker Allowance",
+                      "employmentId"   -> JsNull
+                    )
+                  )
+                )
+              ),
+              "taxBands" -> JsNull
+            )
+          )
+        )
+
+        json.as[Option[TaxOnOtherIncome]](taxOnOtherIncomeReads) mustBe None
+      }
+
+      "non-coded income is present but tax bands income is null" in {
+        val json = Json.obj(
+          "totalLiability" -> Json.obj(
+            "nonSavings" -> Json.obj(
+              "totalIncome" -> Json.obj(
+                "iabdSummaries" -> JsArray(
+                  Seq(
+                    Json.obj(
+                      "amount"         -> 100,
+                      "type"           -> 19,
+                      "npsDescription" -> "Non-Coded Income",
+                      "employmentId"   -> JsNull
+                    ),
+                    Json.obj(
+                      "amount"         -> 100,
+                      "type"           -> 84,
+                      "npsDescription" -> "Job-Seeker Allowance",
+                      "employmentId"   -> JsNull
+                    )
+                  )
+                )
+              ),
+              "taxBands" -> JsArray(
+                Seq(
+                  Json.obj(
+                    "bandType" -> "B",
+                    "income"   -> JsNull,
+                    "taxCode"  -> "BR",
+                    "rate"     -> 20
+                  )
+                )
+              )
+            )
+          )
+        )
+
+        json.as[Option[TaxOnOtherIncome]](taxOnOtherIncomeReads) mustBe None
+      }
+    }
+  }
+
+  "taxOnOtherIncomeRead" must {
+    "return income" when {
+      "non-coded income is present" in {
+        val json = Json.obj(
+          "totalLiability" -> Json.obj(
+            "nonSavings" -> Json.obj(
+              "totalIncome" -> Json.obj(
+                "iabdSummaries" -> JsArray(
+                  Seq(
+                    Json.obj(
+                      "amount"         -> 100,
+                      "type"           -> 19,
+                      "npsDescription" -> "Non-Coded Income",
+                      "employmentId"   -> JsNull
+                    ),
+                    Json.obj(
+                      "amount"         -> 100,
+                      "type"           -> 84,
+                      "npsDescription" -> "Job-Seeker Allowance",
+                      "employmentId"   -> JsNull
+                    )
+                  )
+                )
+              ),
+              "taxBands" -> JsArray(
+                Seq(
+                  Json.obj(
+                    "bandType" -> "B",
+                    "income"   -> 1000,
+                    "taxCode"  -> "BR",
+                    "rate"     -> 40
+                  ),
+                  Json.obj(
+                    "bandType" -> "D0",
+                    "taxCode"  -> "BR",
+                    "income"   -> 1000,
+                    "rate"     -> 20
+                  )
+                )
+              )
+            )
+          )
+        )
+        json.as[Option[BigDecimal]](taxOnOtherIncomeRead) mustBe Some(40)
+      }
+    }
+
+    "return none" when {
+      "details are not present" in {
+        val json = Json.obj(
+          "totalLiability" -> Json.obj(
+            "nonSavings" -> Json.obj(
+              "totalIncome" -> Json.obj(
+                "iabdSummaries" -> JsArray()
+              ),
+              "taxBands" -> JsArray(
+                Seq(
+                  Json.obj(
+                    "bandType" -> "B",
+                    "income"   -> 1000,
+                    "taxCode"  -> "BR",
+                    "rate"     -> 40
+                  ),
+                  Json.obj(
+                    "bandType" -> "D0",
+                    "taxCode"  -> "BR",
+                    "income"   -> 1000,
+                    "rate"     -> 20
+                  )
+                )
+              )
+            )
+          )
+        )
+        json.as[Option[BigDecimal]](taxOnOtherIncomeRead) mustBe None
+      }
+    }
+  }
+
+}
