+++ test/uk/gov/hmrc/tai/model/domain/taxAdjustments/TaxAdjustmentComponentSpec.scala
@@ -14,106 +14,15 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.tai.model.domain.formatters
+package uk.gov.hmrc.tai.model.domain.taxAdjustments
 
 import org.mockito.MockitoSugar
 import org.scalatestplus.play.PlaySpec
-import play.api.libs.json.{JsArray, JsNull, Json}
-import uk.gov.hmrc.tai.model.domain.TaxOnOtherIncome
-import uk.gov.hmrc.tai.model.domain.TaxOnOtherIncome.{taxAccountSummaryReads, taxOnOtherIncomeRead, taxOnOtherIncomeReads}
+import play.api.libs.json.{JsNull, Json}
 import uk.gov.hmrc.tai.model.domain.taxAdjustments.TaxAdjustmentComponent.{alreadyTaxedAtSourceReads, otherTaxDueReads, reliefsGivingBackTaxReads, taxAdjustmentComponentReads, taxReliefFormattersReads}
-import uk.gov.hmrc.tai.model.domain.taxAdjustments._
 
-class TaxAccountSummaryHodFormattersSpec extends PlaySpec with MockitoSugar {
-
-  "taxAccountSummaryReads" must {
-    "return the totalEstTax from the hods response" when {
-      "totalLiability val is present in totalLiability section" in {
-        val json = Json.obj(
-          "totalLiability" -> Json.obj(
-            "totalLiability" -> 1234.56
-          )
-        )
-
-        json.as[BigDecimal](taxAccountSummaryReads) mustBe BigDecimal(1234.56)
-      }
-    }
-    "return zero totalEstTax" when {
-      "totalLiability section is NOT present" in {
-        val json = Json.obj()
-        json.as[BigDecimal](taxAccountSummaryReads) mustBe BigDecimal(0)
-      }
-      "totalLiability section is null" in {
-        val json = Json.obj(
-          "totalLiability" -> JsNull
-        )
-        json.as[BigDecimal](taxAccountSummaryReads) mustBe BigDecimal(0)
-      }
-      "totalLiability section is present but the totalLiability value is not present inside" in {
-        val json = Json.obj(
-          "totalLiability" -> Json.obj()
-        )
-        json.as[BigDecimal](taxAccountSummaryReads) mustBe BigDecimal(0)
-      }
-      "totalLiability section is present but the totalLiability value is null inside" in {
-        val json = Json.obj(
-          "totalLiability" -> JsNull
-        )
-        json.as[BigDecimal](taxAccountSummaryReads) mustBe BigDecimal(0)
-      }
-    }
-
-    "return totalEstTax" when {
-      "tax on other income is present" in {
-        val json = Json.obj(
-          "totalLiability" -> Json.obj(
-            "totalLiability" -> 1234.56,
-            "nonSavings" -> Json.obj(
-              "totalIncome" -> Json.obj(
-                "iabdSummaries" -> JsArray(
-                  Seq(
-                    Json.obj(
-                      "amount"         -> 100,
-                      "type"           -> 19,
-                      "npsDescription" -> "Non-Coded Income",
-                      "employmentId"   -> JsNull
-                    ),
-                    Json.obj(
-                      "amount"         -> 100,
-                      "type"           -> 84,
-                      "npsDescription" -> "Job-Seeker Allowance",
-                      "employmentId"   -> JsNull
-                    )
-                  )
-                )
-              ),
-              "taxBands" -> JsArray(
-                Seq(
-                  Json.obj(
-                    "bandType" -> "B",
-                    "income"   -> 1000,
-                    "taxCode"  -> "BR",
-                    "rate"     -> 40
-                  ),
-                  Json.obj(
-                    "bandType" -> "D0",
-                    "taxCode"  -> "BR",
-                    "income"   -> 1000,
-                    "rate"     -> 20
-                  )
-                )
-              )
-            )
-          )
-        )
-
-        json.as[BigDecimal](taxAccountSummaryReads) mustBe BigDecimal(1194.56)
-      }
-
-    }
-  }
-
-  "ReliefsGivingBackTaxFormatters" must {
+class TaxAdjustmentComponentSpec extends PlaySpec with MockitoSugar {
+  "reliefsGivingBackTaxReads" must {
     "return empty components" when {
       "relief giving back tax is null" in {
         val json = Json.obj(
@@ -188,7 +97,7 @@ class TaxAccountSummaryHodFormattersSpec extends PlaySpec with MockitoSugar {
     }
   }
 
-  "OtherTaxDueFormatters" must {
+  "otherTaxDueReads" must {
     "return none" when {
       "other tax due is null" in {
         val json = Json.obj(
@@ -258,7 +167,7 @@ class TaxAccountSummaryHodFormattersSpec extends PlaySpec with MockitoSugar {
     }
   }
 
-  "AlreadyTaxedAtSourceFormatters" must {
+  "alreadyTaxedAtSourceReads" must {
     "return none" when {
       "already taxed details is null" in {
         val json = Json.obj(
@@ -327,293 +236,7 @@ class TaxAccountSummaryHodFormattersSpec extends PlaySpec with MockitoSugar {
     }
   }
 
-  "TaxOnOtherIncomeFormatters" must {
-    "return tax on other income" when {
-      "non-coded income is present and highest rate is 40%" in {
-        val json = Json.obj(
-          "totalLiability" -> Json.obj(
-            "nonSavings" -> Json.obj(
-              "totalIncome" -> Json.obj(
-                "iabdSummaries" -> JsArray(
-                  Seq(
-                    Json.obj(
-                      "amount"         -> 100,
-                      "type"           -> 19,
-                      "npsDescription" -> "Non-Coded Income",
-                      "employmentId"   -> JsNull
-                    ),
-                    Json.obj(
-                      "amount"         -> 100,
-                      "type"           -> 84,
-                      "npsDescription" -> "Job-Seeker Allowance",
-                      "employmentId"   -> JsNull
-                    )
-                  )
-                )
-              ),
-              "taxBands" -> JsArray(
-                Seq(
-                  Json.obj(
-                    "bandType" -> "B",
-                    "income"   -> 1000,
-                    "taxCode"  -> "BR",
-                    "rate"     -> 40
-                  ),
-                  Json.obj(
-                    "bandType" -> "D0",
-                    "taxCode"  -> "BR",
-                    "income"   -> 1000,
-                    "rate"     -> 20
-                  )
-                )
-              )
-            )
-          )
-        )
-
-        json.as[Option[TaxOnOtherIncome]](taxOnOtherIncomeReads) mustBe Some(TaxOnOtherIncome(40))
-      }
-
-      "non-coded income is present and equal to highest rate income " in {
-        val json = Json.obj(
-          "totalLiability" -> Json.obj(
-            "nonSavings" -> Json.obj(
-              "totalIncome" -> Json.obj(
-                "iabdSummaries" -> JsArray(
-                  Seq(
-                    Json.obj(
-                      "amount"         -> 1000,
-                      "type"           -> 19,
-                      "npsDescription" -> "Non-Coded Income",
-                      "employmentId"   -> JsNull
-                    ),
-                    Json.obj(
-                      "amount"         -> 100,
-                      "type"           -> 84,
-                      "npsDescription" -> "Job-Seeker Allowance",
-                      "employmentId"   -> JsNull
-                    )
-                  )
-                )
-              ),
-              "taxBands" -> JsArray(
-                Seq(
-                  Json.obj(
-                    "bandType" -> "B",
-                    "income"   -> 1000,
-                    "taxCode"  -> "BR",
-                    "rate"     -> 40
-                  ),
-                  Json.obj(
-                    "bandType" -> "D0",
-                    "taxCode"  -> "BR",
-                    "income"   -> 1000,
-                    "rate"     -> 20
-                  )
-                )
-              )
-            )
-          )
-        )
-
-        json.as[Option[TaxOnOtherIncome]](taxOnOtherIncomeReads) mustBe Some(TaxOnOtherIncome(400))
-      }
-
-      "non-coded income is present and scattered in multiple rate bands " in {
-        val json = Json.obj(
-          "totalLiability" -> Json.obj(
-            "nonSavings" -> Json.obj(
-              "totalIncome" -> Json.obj(
-                "iabdSummaries" -> JsArray(
-                  Seq(
-                    Json.obj(
-                      "amount"         -> 10000,
-                      "type"           -> 19,
-                      "npsDescription" -> "Non-Coded Income",
-                      "employmentId"   -> JsNull
-                    ),
-                    Json.obj(
-                      "amount"         -> 100,
-                      "type"           -> 84,
-                      "npsDescription" -> "Job-Seeker Allowance",
-                      "employmentId"   -> JsNull
-                    )
-                  )
-                )
-              ),
-              "taxBands" -> JsArray(
-                Seq(
-                  Json.obj(
-                    "bandType" -> "B",
-                    "income"   -> 5000,
-                    "taxCode"  -> "BR",
-                    "rate"     -> 40
-                  ),
-                  Json.obj(
-                    "bandType" -> "D0",
-                    "taxCode"  -> "BR",
-                    "income"   -> 1000,
-                    "rate"     -> 20
-                  ),
-                  Json.obj(
-                    "bandType" -> "D0",
-                    "taxCode"  -> "BR",
-                    "income"   -> 8000,
-                    "rate"     -> 10
-                  )
-                )
-              )
-            )
-          )
-        )
-
-        json.as[Option[TaxOnOtherIncome]](taxOnOtherIncomeReads) mustBe Some(TaxOnOtherIncome(2600))
-      }
-
-      "non-coded income is present and highest rate is 20%" in {
-        val json = Json.obj(
-          "totalLiability" -> Json.obj(
-            "nonSavings" -> Json.obj(
-              "totalIncome" -> Json.obj(
-                "iabdSummaries" -> JsArray(
-                  Seq(
-                    Json.obj(
-                      "amount"         -> 100,
-                      "type"           -> 19,
-                      "npsDescription" -> "Non-Coded Income",
-                      "employmentId"   -> JsNull
-                    ),
-                    Json.obj(
-                      "amount"         -> 100,
-                      "type"           -> 84,
-                      "npsDescription" -> "Job-Seeker Allowance",
-                      "employmentId"   -> JsNull
-                    )
-                  )
-                )
-              ),
-              "taxBands" -> JsArray(
-                Seq(
-                  Json.obj(
-                    "bandType" -> "B",
-                    "income"   -> 1000,
-                    "taxCode"  -> "BR",
-                    "rate"     -> 20
-                  )
-                )
-              )
-            )
-          )
-        )
-
-        json.as[Option[TaxOnOtherIncome]](taxOnOtherIncomeReads) mustBe Some(TaxOnOtherIncome(20))
-      }
-
-      "non-coded income is not present" in {
-        val json = Json.obj(
-          "totalLiability" -> Json.obj(
-            "nonSavings" -> Json.obj(
-              "totalIncome" -> Json.obj(
-                "iabdSummaries" -> JsArray(
-                  Seq(
-                    Json.obj(
-                      "amount"         -> 100,
-                      "type"           -> 84,
-                      "npsDescription" -> "Job-Seeker Allowance",
-                      "employmentId"   -> JsNull
-                    )
-                  )
-                )
-              ),
-              "taxBands" -> JsArray(
-                Seq(
-                  Json.obj(
-                    "bandType" -> "B",
-                    "income"   -> 1000,
-                    "taxCode"  -> "BR",
-                    "rate"     -> 20
-                  )
-                )
-              )
-            )
-          )
-        )
-
-        json.as[Option[TaxOnOtherIncome]](taxOnOtherIncomeReads) mustBe None
-      }
-
-      "non-coded income is present and tax bands are not present" in {
-        val json = Json.obj(
-          "totalLiability" -> Json.obj(
-            "nonSavings" -> Json.obj(
-              "totalIncome" -> Json.obj(
-                "iabdSummaries" -> JsArray(
-                  Seq(
-                    Json.obj(
-                      "amount"         -> 100,
-                      "type"           -> 19,
-                      "npsDescription" -> "Non-Coded Income",
-                      "employmentId"   -> JsNull
-                    ),
-                    Json.obj(
-                      "amount"         -> 100,
-                      "type"           -> 84,
-                      "npsDescription" -> "Job-Seeker Allowance",
-                      "employmentId"   -> JsNull
-                    )
-                  )
-                )
-              ),
-              "taxBands" -> JsNull
-            )
-          )
-        )
-
-        json.as[Option[TaxOnOtherIncome]](taxOnOtherIncomeReads) mustBe None
-      }
-
-      "non-coded income is present but tax bands income is null" in {
-        val json = Json.obj(
-          "totalLiability" -> Json.obj(
-            "nonSavings" -> Json.obj(
-              "totalIncome" -> Json.obj(
-                "iabdSummaries" -> JsArray(
-                  Seq(
-                    Json.obj(
-                      "amount"         -> 100,
-                      "type"           -> 19,
-                      "npsDescription" -> "Non-Coded Income",
-                      "employmentId"   -> JsNull
-                    ),
-                    Json.obj(
-                      "amount"         -> 100,
-                      "type"           -> 84,
-                      "npsDescription" -> "Job-Seeker Allowance",
-                      "employmentId"   -> JsNull
-                    )
-                  )
-                )
-              ),
-              "taxBands" -> JsArray(
-                Seq(
-                  Json.obj(
-                    "bandType" -> "B",
-                    "income"   -> JsNull,
-                    "taxCode"  -> "BR",
-                    "rate"     -> 20
-                  )
-                )
-              )
-            )
-          )
-        )
-
-        json.as[Option[TaxOnOtherIncome]](taxOnOtherIncomeReads) mustBe None
-      }
-    }
-  }
-
-  "LiabilityComponentReads" must {
+  "taxAdjustmentComponentReads" must {
     "return liability components" when {
       "only reliefs are present" in {
         val json = Json.obj(
@@ -729,86 +352,7 @@ class TaxAccountSummaryHodFormattersSpec extends PlaySpec with MockitoSugar {
     }
   }
 
-  "taxOnOtherIncomeRead" must {
-    "return income" when {
-      "non-coded income is present" in {
-        val json = Json.obj(
-          "totalLiability" -> Json.obj(
-            "nonSavings" -> Json.obj(
-              "totalIncome" -> Json.obj(
-                "iabdSummaries" -> JsArray(
-                  Seq(
-                    Json.obj(
-                      "amount"         -> 100,
-                      "type"           -> 19,
-                      "npsDescription" -> "Non-Coded Income",
-                      "employmentId"   -> JsNull
-                    ),
-                    Json.obj(
-                      "amount"         -> 100,
-                      "type"           -> 84,
-                      "npsDescription" -> "Job-Seeker Allowance",
-                      "employmentId"   -> JsNull
-                    )
-                  )
-                )
-              ),
-              "taxBands" -> JsArray(
-                Seq(
-                  Json.obj(
-                    "bandType" -> "B",
-                    "income"   -> 1000,
-                    "taxCode"  -> "BR",
-                    "rate"     -> 40
-                  ),
-                  Json.obj(
-                    "bandType" -> "D0",
-                    "taxCode"  -> "BR",
-                    "income"   -> 1000,
-                    "rate"     -> 20
-                  )
-                )
-              )
-            )
-          )
-        )
-        json.as[Option[BigDecimal]](taxOnOtherIncomeRead) mustBe Some(40)
-      }
-    }
-
-    "return none" when {
-      "details are not present" in {
-        val json = Json.obj(
-          "totalLiability" -> Json.obj(
-            "nonSavings" -> Json.obj(
-              "totalIncome" -> Json.obj(
-                "iabdSummaries" -> JsArray()
-              ),
-              "taxBands" -> JsArray(
-                Seq(
-                  Json.obj(
-                    "bandType" -> "B",
-                    "income"   -> 1000,
-                    "taxCode"  -> "BR",
-                    "rate"     -> 40
-                  ),
-                  Json.obj(
-                    "bandType" -> "D0",
-                    "taxCode"  -> "BR",
-                    "income"   -> 1000,
-                    "rate"     -> 20
-                  )
-                )
-              )
-            )
-          )
-        )
-        json.as[Option[BigDecimal]](taxOnOtherIncomeRead) mustBe None
-      }
-    }
-  }
-
-  "taxReliefComponentRead" must {
+  "taxReliefFormattersReads" must {
     "return empty sequence" when {
       "basic rate extension is null" in {
         val json = Json.obj(
